--- src/components/Movies/MoviesCard/CardActions/SavedMovie/SavedMovie.js
@@ -1,14 +1,0 @@
-import React from 'react';
-import { saveIcons } from '../../../../../utils/constants';
-
-function SavedMovie() {
-  return (
-    <img
-      className="savedMovie"
-      alt="Сохранен"
-      src={saveIcons}
-    />
-  );
-}
-
-export default SavedMovie;
