--- src/components/Movies/MoviesCard/MoviesCard.js
@@ -1,351 +1,0 @@
-import React from 'react';
-import testCardImg from '../../../images/card33Slova.jpeg';
-import SaveMovie from './CardActions/SaveMovie/SaveMovie';
-import SavedMovie from './CardActions/SavedMovie/SavedMovie';
-
-function MoviesCard() {
-  return (
-    <div className="moviesCard">
-      <div className="moviesCard__main">
-        <a
-          href="https://ya.ru"
-          className="moviesCard__link"
-          target="_blank"
-          rel="noreferrer"
-        >
-          <img
-            className="moviesCard__img"
-            alt="33 слова о дизайне"
-            src={testCardImg}
-            id="cardImage"
-          />
-        </a>
-        <div className="moviesCard__action-block">
-          <SaveMovie />
-        </div>
-        <h3 className="moviesCard__title">
-          33 слова о дизайне
-        </h3>
-        <div className="moviesCard__duration">
-          <h3 className="moviesCard__time">1ч 17м</h3>
-        </div>
-      </div>
-      <div className="moviesCard__main">
-        <a
-          href="https://ya.ru"
-          className="moviesCard__link"
-          target="_blank"
-          rel="noreferrer"
-        >
-          <img
-            className="moviesCard__img"
-            alt="33 слова о дизайне"
-            src={testCardImg}
-            id="cardImage"
-          />
-        </a>
-        <div className="moviesCard__action-block">
-          <SavedMovie />
-        </div>
-        <h3 className="moviesCard__title">
-          33 слова о дизайне
-        </h3>
-        <div className="moviesCard__duration">
-          <h3 className="moviesCard__time">1ч 17м</h3>
-        </div>
-      </div>
-      <div className="moviesCard__main">
-        <a
-          href="https://ya.ru"
-          className="moviesCard__link"
-          target="_blank"
-          rel="noreferrer"
-        >
-          <img
-            className="moviesCard__img"
-            alt="33 слова о дизайне"
-            src={testCardImg}
-            id="cardImage"
-          />
-        </a>
-        <div className="moviesCard__action-block">
-          <button
-            className="moviesCard__action hover"
-            type="button"
-          >
-            Сохранить
-          </button>
-        </div>
-        <h3 className="moviesCard__title">
-          33 слова о дизайне
-        </h3>
-        <div className="moviesCard__duration">
-          <h3 className="moviesCard__time">1ч 17м</h3>
-        </div>
-      </div>
-      <div className="moviesCard__main">
-        <a
-          href="https://ya.ru"
-          className="moviesCard__link"
-          target="_blank"
-          rel="noreferrer"
-        >
-          <img
-            className="moviesCard__img"
-            alt="33 слова о дизайне"
-            src={testCardImg}
-            id="cardImage"
-          />
-        </a>
-        <div className="moviesCard__action-block">
-          <button
-            className="moviesCard__action hover"
-            type="button"
-          >
-            Сохранить
-          </button>
-        </div>
-        <h3 className="moviesCard__title">
-          33 слова о дизайне
-        </h3>
-        <div className="moviesCard__duration">
-          <h3 className="moviesCard__time">1ч 17м</h3>
-        </div>
-      </div>
-      <div className="moviesCard__main">
-        <a
-          href="https://ya.ru"
-          className="moviesCard__link"
-          target="_blank"
-          rel="noreferrer"
-        >
-          <img
-            className="moviesCard__img"
-            alt="33 слова о дизайне"
-            src={testCardImg}
-            id="cardImage"
-          />
-        </a>
-        <div className="moviesCard__action-block">
-          <button
-            className="moviesCard__action hover"
-            type="button"
-          >
-            Сохранить
-          </button>
-        </div>
-        <h3 className="moviesCard__title">
-          33 слова о дизайне
-        </h3>
-        <div className="moviesCard__duration">
-          <h3 className="moviesCard__time">1ч 17м</h3>
-        </div>
-      </div>
-      <div className="moviesCard__main">
-        <a
-          href="https://ya.ru"
-          className="moviesCard__link"
-          target="_blank"
-          rel="noreferrer"
-        >
-          <img
-            className="moviesCard__img"
-            alt="33 слова о дизайне"
-            src={testCardImg}
-            id="cardImage"
-          />
-        </a>
-        <div className="moviesCard__action-block">
-          <button
-            className="moviesCard__action hover"
-            type="button"
-          >
-            Сохранить
-          </button>
-        </div>
-        <h3 className="moviesCard__title">
-          33 слова о дизайне
-        </h3>
-        <div className="moviesCard__duration">
-          <h3 className="moviesCard__time">1ч 17м</h3>
-        </div>
-      </div>
-      <div className="moviesCard__main">
-        <a
-          href="https://ya.ru"
-          className="moviesCard__link"
-          target="_blank"
-          rel="noreferrer"
-        >
-          <img
-            className="moviesCard__img"
-            alt="33 слова о дизайне"
-            src={testCardImg}
-            id="cardImage"
-          />
-        </a>
-        <div className="moviesCard__action-block">
-          <button
-            className="moviesCard__action hover"
-            type="button"
-          >
-            Сохранить
-          </button>
-        </div>
-        <h3 className="moviesCard__title">
-          33 слова о дизайне
-        </h3>
-        <div className="moviesCard__duration">
-          <h3 className="moviesCard__time">1ч 17м</h3>
-        </div>
-      </div>
-      <div className="moviesCard__main">
-        <a
-          href="https://ya.ru"
-          className="moviesCard__link"
-          target="_blank"
-          rel="noreferrer"
-        >
-          <img
-            className="moviesCard__img"
-            alt="33 слова о дизайне"
-            src={testCardImg}
-            id="cardImage"
-          />
-        </a>
-        <div className="moviesCard__action-block">
-          <button
-            className="moviesCard__action hover"
-            type="button"
-          >
-            Сохранить
-          </button>
-        </div>
-        <h3 className="moviesCard__title">
-          33 слова о дизайне
-        </h3>
-        <div className="moviesCard__duration">
-          <h3 className="moviesCard__time">1ч 17м</h3>
-        </div>
-      </div>
-      <div className="moviesCard__main">
-        <a
-          href="https://ya.ru"
-          className="moviesCard__link"
-          target="_blank"
-          rel="noreferrer"
-        >
-          <img
-            className="moviesCard__img"
-            alt="33 слова о дизайне"
-            src={testCardImg}
-            id="cardImage"
-          />
-        </a>
-        <div className="moviesCard__action-block">
-          <button
-            className="moviesCard__action hover"
-            type="button"
-          >
-            Сохранить
-          </button>
-        </div>
-        <h3 className="moviesCard__title">
-          33 слова о дизайне
-        </h3>
-        <div className="moviesCard__duration">
-          <h3 className="moviesCard__time">1ч 17м</h3>
-        </div>
-      </div>
-      <div className="moviesCard__main">
-        <a
-          href="https://ya.ru"
-          className="moviesCard__link"
-          target="_blank"
-          rel="noreferrer"
-        >
-          <img
-            className="moviesCard__img"
-            alt="33 слова о дизайне"
-            src={testCardImg}
-            id="cardImage"
-          />
-        </a>
-        <div className="moviesCard__action-block">
-          <button
-            className="moviesCard__action hover"
-            type="button"
-          >
-            Сохранить
-          </button>
-        </div>
-        <h3 className="moviesCard__title">
-          33 слова о дизайне
-        </h3>
-        <div className="moviesCard__duration">
-          <h3 className="moviesCard__time">1ч 17м</h3>
-        </div>
-      </div>
-      <div className="moviesCard__main">
-        <a
-          href="https://ya.ru"
-          className="moviesCard__link"
-          target="_blank"
-          rel="noreferrer"
-        >
-          <img
-            className="moviesCard__img"
-            alt="33 слова о дизайне"
-            src={testCardImg}
-            id="cardImage"
-          />
-        </a>
-        <div className="moviesCard__action-block">
-          <button
-            className="moviesCard__action hover"
-            type="button"
-          >
-            Сохранить
-          </button>
-        </div>
-        <h3 className="moviesCard__title">
-          33 слова о дизайне
-        </h3>
-        <div className="moviesCard__duration">
-          <h3 className="moviesCard__time">1ч 17м</h3>
-        </div>
-      </div>
-      <div className="moviesCard__main">
-        <a
-          href="https://ya.ru"
-          className="moviesCard__link"
-          target="_blank"
-          rel="noreferrer"
-        >
-          <img
-            className="moviesCard__img"
-            alt="33 слова о дизайне"
-            src={testCardImg}
-            id="cardImage"
-          />
-        </a>
-        <div className="moviesCard__action-block">
-          <button
-            className="moviesCard__action hover"
-            type="button"
-          >
-            Сохранить
-          </button>
-        </div>
-        <h3 className="moviesCard__title">
-          33 слова о дизайне
-        </h3>
-        <div className="moviesCard__duration">
-          <h3 className="moviesCard__time">1ч 17м</h3>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-export default MoviesCard;
