+++ src/components/MoviesCard/CardActions/SavedIcon/SavedIcon.css
@@ -1,4 +1,4 @@
-.savedMovie {
+.saved-icon {
   width: 21px;
   height: 21px;
 }
