+++ src/components/MoviesCard/CardActions/deleteMovieIcon/deleteMovieIcon.css
@@ -1,0 +1,16 @@
+.delete-movie {
+  border: none;
+  cursor: pointer;
+  margin: 0;
+  padding: 0;
+  width: 21px;
+  height: 21px;
+  background: transparent;
+}
+
+.delete-movie__icon {
+  margin: 0;
+  padding: 0;
+  width: 21px;
+  height: 21px;
+}
