+++ src/components/SavedMovies/SavedMovies.js
@@ -1,17 +1,16 @@
 import React from 'react';
 import SearchForm from '../SearchForm/SearchForm';
 import MoviesCardList from '../MoviesCardList/MoviesCardList';
-
 import Preloader from '../Preloader/Preloader';
 
-function Movies({ isShortMovie, isShortMovieHandler }) {
+function SavedMovies() {
   return (
-    <div className="movies">
-      <SearchForm isShortMovie={isShortMovie} isShortMovieHandler={isShortMovieHandler} />
+    <div className="SavedMovies">
+      <SearchForm />
       <MoviesCardList />
       <Preloader />
     </div>
   );
 }
 
-export default Movies;
+export default SavedMovies;
